SystemCode/backend/app/dataservice/sql_api/requirements.txt
geoalchemy2==0.17.0
pydantic==2.12.3
python-dotenv==1.1.1
SQLAlchemy==2.0.44
asyncpg==0.29.0
greenlet==3.2.4
